Added Initial ado files.
These files come directly from the private subversion repository
that we've been using for the MESA L-SLIS project. (For my
reference: revision numer 1460, the deliverable branch.)

There are a couple of help files, but I think the one for stbl is
out of date. I intend to add help files in the future.